Add debounce option to prevent the reload event from triggering uncessarily. Solves #134
I added a `debounce` option that when used turns the change callback function into a debounced function a la Lodash that will wait `options.debounce` milliseconds since the last call to actually execute. If the `debounce` option isn't provided or is 0, then it behave as it did before this change.