Foxy cookies.stripDomain option configurable
The cookies.stripDomain option is enabled by default in Foxy.
The problem with that is that there is no way of preventing the proxy server from altering your cookies. In some cases altering the cookies may break the session, like in this case: https://github.com/shakyShane/foxy/pull/6

It's anyway good citizenship to have the option to opt-out from features that may corrupt the request headers in edge cases like that.