fix: https.passphrase option should be respected
serverUtils.getHttpsOptions() was ignoring the `https.passphrase` options since the "default" https options were merged on top of user options. Another solution to this is to just omit the `passphrase` key entirely in the defaults section, but I took this approach in order to preserve the current behavior where it's set to an empty string by default.